General restructuring of entities
All Users (CorporateUser and PersonalUser) store two HashMaps that provide a local copy of their relationships to other Users. Each hashmap corresponds to a different type of related user, either other PersonalUser or CorporateUser. These classes have many common methods for modifying their connections lists, so a new class Connections was created to provide an algorithmic framework for implementing these relationships.

The functionality of Client and Company have been migrated to User and Organization, respectively. They are ready to be deleted from the entities. Currently, just waiting on refactoring of other classes that still presently depend on Client and Company.